pyproject.toml
[tool.poetry]
name = "technical-assessment-2024"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.12.0"
django = "^5.0"
djangorestframework = "^3.15.1"
django-autoslug = "^1.9.8"
django-filter = "21.1"
drf-yasg = "^1.21.7"
celery = "^5.3.6"
python-dotenv = "^1.0.1"
redis = "^5.0.3"
django-celery-results = "^2.5.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.0.0"
pillow = "^10.3.0"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
psycopg2 = "^2.9.3"
gunicorn = "^21.1.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "^7.4.4"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
